Job-Portal-Backend/requirements.txt
asgiref==3.5.2
astroid==2.3.3
backports.zoneinfo==0.2.1
certifi==2019.11.28
chardet==3.0.4
Django==4.0
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==2.8
importlib-metadata==4.12.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.4.1
mccabe==0.6.1
Pillow==9.1.1
PyJWT==2.6.0
pytz==2022.1
requests==2.22.0
six==1.13.0
sqlparse==0.4.2
sslcommerz-python==0.0.7
typed-ast==1.4.0
tzdata==2022.1
urllib3==1.25.7
wrapt==1.11.2
zipp==3.8.1
